proxyconn: add an option to forward unclaimed CONNECT requests
When the new ForwardConnect option is set true, the Bridge will not reject
CONNECT requests for hosts not listed in Allow, but instead will open direct
connections to them and splice the connections together.

This permits the Bridge to act as a proxy for ANY valid target, while only
capturing the "interesting" hosts for the proxy.